jira_consumer/requirements.txt
jira==3.8.0
python-dotenv==1.0.1
gunicorn==23.0.0
sqs_consumer @ git+https://github.com/RoAsCro/SQSConsumer